fastapi/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.31.1"}
debugpy = "^1.8.7"
requests = "^2.32.3"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
